Add value type template parameter to Reader and Writer
