--- src/MSArray.cc
@@ -1,7 +1,0 @@
-/* -*- mode: c++; c-basic-offset: 2; indent-tabs-mode: nil; -*- */
-#include <MSArray.h>
-
-void
-mpims::swap(MSArray& array1, MSArray& array2) {
-  array1.swap(array2);
-}
